+++ safari-app/SmoothScroll GPT/SmoothScroll GPT/AppDelegate.swift
@@ -1,0 +1,21 @@
+//
+//  AppDelegate.swift
+//  SmoothScroll GPT
+//
+//  Created by shuitata on 2026/2/24.
+//
+
+import Cocoa
+
+@main
+class AppDelegate: NSObject, NSApplicationDelegate {
+
+    func applicationDidFinishLaunching(_ notification: Notification) {
+        // Override point for customization after application launch.
+    }
+
+    func applicationShouldTerminateAfterLastWindowClosed(_ sender: NSApplication) -> Bool {
+        return true
+    }
+
+}